// UnrealEngine52/Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export enum Priority {

	// Lowest priority
	Lowest = "Lowest",

	// Below normal priority
	BelowNormal = "BelowNormal",

	// Normal priority
	Normal = "Normal",

	// Above normal priority
	AboveNormal = "AboveNormal",

	// High priority
	High = "High",

	// Highest priority
	Highest = "Highest"
}
// ...
export enum JobStepOutcome {

	// Finished with errors
	Failure = "Failure",

	// Finished with warnings
	Warnings = "Warnings",

	// Finished Succesfully
	Success = "Success",

	// Unspecified, skipped, aborted, etc
	Unspecified = "Unspecified"
}
// ...
export type ProjectData = GetProjectResponse & {
	streams?: StreamData[];
}

export type StreamData = GetStreamResponse & {
	project?: ProjectData;
	// full path as returned by server
	fullname?: string;

}

export type JobData = GetJobResponse & {
	graphRef?: GetGraphResponse;
}
// ...
export type JobStreamQuery = {
	filter?: string;
	template?: string[];
	preflightStartedByUserId?: string,
	includePreflight?: boolean,
	maxCreateTime?: string;
	modifiedAfter?: string;
	index?: number;
	count?: number;
	consistentRead?: boolean;

};
// ...
export type GetReportResponse = {

	// Report placement
	placement: ReportPlacement;

	/** Name of the report */
	name: string;

	/** The artifact id */
	artifactId: string;
}


/**Information about a job */
export type GetJobResponse = {

	/**Unique Id for the job */
	id: string;

	/**Unique id of the stream containing this job */
	streamId: string;

	/**Name of the job */
	name: string;

	/** The changelist number to build */
	change?: number;

	/** The preflight changelist number */
	preflightChange?: number;

	/** Description of the preflight */
	preflightDescription?: string;

	/** The template type */
	templateId?: string;

	/** Hash of the actual template data */
	templateHash?: string;

	/** Hash of the graph for this job */
	graphHash?: string;

	/** The user that started this job */
	startedByUserInfo?: GetThinUserInfoResponse;

	/** The user that started this job */
	abortedByUserInfo?: GetThinUserInfoResponse;

	/** The roles to impersonate when executing this job */
	roles: string[];

	/** Priority of the job */
	priority: Priority;

	/** Whether the change will automatically be submitted or not */
	autoSubmit?: boolean;

	/** The submitted changelist number */
	autoSubmitChange?: number;

	/**  Message produced by trying to auto-submit the change */
	autoSubmitMessage?: string;

	/**Time that the job was created */
	createTime: Date | string;

	/** The global job state */
	state: JobState;

	/** Array of jobstep batches */
	batches?: GetBatchResponse[];

	/**  List of labels */
	labels?: GetLabelStateResponse[];

	/** The default label, containing the state of all steps that are otherwise not matched. */
	defaultLabel?: GetDefaultLabelStateResponse;

	/** List of reports */
	reports?: GetReportResponse[];

	/**  Parameters for the job */
	arguments: string[];

	/**The last update time for this job*/
	updateTime: Date | string;

	/** Whether to update issues based on the outcome of this job */
	updateIssues?: boolean;

	/**  Custom permissions for this object */
	acl?: GetAclResponse;

}
// ...
export type GetBatchResponse = {

	/**Unique id for this batch */
	id: string;

	/**The unique log file id */
	logId?: string;

	/**Index of the group being executed */
	groupIdx: number;

	/**The state of this batch */
	state: JobStepBatchState;

	/**Error code for this batch */
	error: JobStepBatchError;

	/**Steps within this run */
	steps: GetStepResponse[];

	/**The agent assigned to execute this group */
	agentId?: string;

	/** The USD rate of an agent hour */
	agentRate?: number;

	/**The agent session holding this lease */
	sessionId?: string;

	/**The lease that's executing this group */
	leaseId?: string;

	/**The priority of this batch */
	weightedPriority: number;

	/**Time at which the group started (UTC). */
	startTime?: Date | string;

	/**Time at which the group finished (UTC) */
	finishTime?: Date | string;

}
// ...
export type GetGroupResponse = {

	/**The executor to use for this group */
	executor: string;

	/**The type of agent to execute this group */
	agentType: string;

	/**Nodes in the group */
	nodes: GetNodeResponse[];
}
// ...
export type GetProjectResponse = {
	/**Unique id of the project */
	id: string;

	/**Name of the project */
	name: string;

	/**Order to display this project on the dashboard */
	order: number;

	/**List of streams that are in this project */
	// streams?: ProjectStreamData[];

	/**List of stream categories to display */
	categories?: GetProjectCategoryResponse[];

	/**Properties for this project. */
	properties: { [key: string]: string };

	/**Custom permissions for this object */
	acl?: GetAclResponse;
}
// ...
export type GetAgentTypeResponse = {

	/**Pool of agents to use for this agent type */
	pool: string;

	/**Name of the workspace to sync */
	workspace?: string;

	/**Pool of agents to use to execute this work */
	requirements?: GetAgentRequirementsResponse;

	/**Path to the temporary storage dir */
	tempStorageDir?: string;

	/**Environment variables to be set when executing the job */
	environment?: { [key: string]: string };

}

/**Information about a workspace type */
export type GetWorkspaceTypeResponse = {

	/**The Perforce server and port (eg. perforce:1666) */
	serverAndPort?: string;

	/**User to log into Perforce with (defaults to buildmachine) */
	userName?: string;

	/**Identifier to distinguish this workspace from other workspaces. Defaults to the workspace type name. */
	identifier?: string;

	/**Override for the stream to sync */
	stream?: string;

	/**Custom view for the workspace */
	view?: string[];

	/**Whether to use an incrementally synced workspace */
	incremental: boolean;
}
// ...
export type GetTemplateRefResponse = GetTemplateResponseBase & {

	/**Unique id of this template ref, (sanitized name) */
	id: string;

	/**Hash of the template definition */
	hash: string;

	/// The schedule for this ref
	schedule?: GetScheduleResponse;

	/** Step state for template in stream */
	stepStates?: GetTemplateStepStateResponse[];

	/** List of queries for the default changelist */
	defaultChange?: ChangeQueryConfig[];

}

/** Specifies defaults for running a preflight */
export type DefaultPreflightConfig = {

	/** The template id to query */
	templateId?: string;

	change?: ChangeQueryConfig;
}


/**  Configuration for an issue workflow */
export type WorkflowConfig = {

	/** Identifier for this workflow */
	id: string;

	/** Name of the tab to post summary data to */
	summaryTab?: string;

}

/**Response describing a stream */
export type GetStreamResponse = {

	/**Unique id of the stream */
	id: string;

	/**Unique id of the project containing this stream */
	projectId: string;

	/**Name of the stream */
	name: string;

	/**The config file path on the server*/
	configPath?: string;

	/**Revision of the config file */
	configRevision?: string;

	/**List of tabs to display for this stream*/
	tabs: GetStreamTabResponse[];

	/** Default template for running preflights */
	defaultPreflight?: DefaultPreflightConfig;

	/**Map of agent name to type */
	agentTypes: { [key: string]: GetAgentTypeResponse };

	/**Map of workspace name to type */
	workspaceTypes?: { [key: string]: GetWorkspaceTypeResponse };

	/**Templates for jobs in this stream */
	templates: GetTemplateRefResponse[];

	/**Properties for this stream */
	properties: { [key: string]: string };

	/** Workflows for this stream */
	workflows: WorkflowConfig[];

	/** Custom permissions for this object */
	acl?: GetAclResponse;

}
// ...
export type GetAclResponse = {
	/**Entries to replace the existing ACL */
	entries: GetAclEntryResponse[];

	/**Whether to inherit permissions from the parent entity */
	inheritPermissions: boolean;
}
// ...
export type GetStreamTabResponse = {

	/**Title of this page */
	title: string;

	type: TabType;

};
// ...
export enum JobState {

	/** Waiting for resources*/
	Waiting = "Waiting",

	/** Currently running one or more steps*/
	Running = "Running",

	/** All steps have completed */
	Complete = "Complete"
}


/**Information about a label */
export type GetLabelResponse = {

	/**Category of the aggregate */
	category: string;

	/**Label for this aggregate */
	name: string;

	/**Label for this aggregate, currently mapped to name property on server */
	dashboardName?: string;

	/**Name to show for this label in UGS */
	ugsName?: string;

	/** Project to display this label for in UGS */
	ugsProject?: string;

	/**Nodes which must be part of the job for the aggregate to be shown */
	requiredNodes: string[];

	/**Nodes to include in the status of this aggregate, if present in the job */
	includedNodes: string[];
}


/**Information about an aggregate */
export type GetAggregateResponse = {

	/**Name of the aggregate */
	name: string;

	/**Nodes which must be part of the job for the aggregate to be shown */
	nodes: string[];

}

/**Information about a graph */
export type GetGraphResponse = {

	/**The hash of the graph */
	hash: string;

	/**Array of nodes for this job */
	groups?: GetGroupResponse[];

	/**List of aggregates */
	namedAggregates?: GetAggregateResponse[];

	/**List of labels for the graph */
	labels?: GetLabelResponse[];

}
// ...
export type GetLabelStateResponse = {
	/**State of the label */
	state?: LabelState;

	/**Outcome of the label */
	outcome?: LabelOutcome;
}

/**Information about the default label (ie. with inlined list of nodes) */
export type GetDefaultLabelStateResponse = GetLabelStateResponse &
{
	/**List of nodes covered by default label */
	nodes: string[];
}
// ...
export type GetThinUserInfoResponse = {

	/**  Id of the user */
	id: string;

	/** Name of the user */
	name: string;

	/**  The user's email address */
	email: string;

}